Adds CI support with Travis
From PR #24.

Fixes #18: Installs MarkLogic and the other dependencies in the Travis container and runs the tests. It’s currently brittle because I’m developing on OS X and Travis runs on Ubuntu. (Completely untested on Windows.) For example, writing to `$ML_HOME/Modules` requires `root` on Linux. `sudo` doesn't know about `nvm` and thus `npm`, etc.

There may be future opportunities to improve this with Docker on Travis (#23).